+++ src/biz/bokhorst/xprivacy/PrivacyManager.java
@@ -136,6 +136,7 @@ public class PrivacyManager {
 	public final static String cSettingAOSPMode = "AOSPMode";
 	public final static String cSettingChangelog = "Changelog";
 	public final static String cSettingUpdates = "Updates";
+	public final static String cSettingMethodExpert = "MethodExpert";
 
 	public final static String cSettingODExpert = "ODExpert";
 	public final static String cSettingODCategory = "ODCategory";