+++ CHANGELOG.md
@@ -26,8 +26,9 @@ See for more information about XPrivacy 3 [this FAQ](https://github.com/M66B/XPr
 **Next release**
 
 * Fixed all problems reported through the debug info
-* Showing message when expert mode enabled
+* Showing message when enabling expert mode / using an expert function
 * Scroll view for toggle restrictions radio buttons
+* Updated Dutch translation
 * Updated Indonesian translation
 
 [Open issues](https://github.com/M66B/XPrivacy/issues?state=open)